Use double precision numbers

+++ examples/gen_2d.rs
@@ -29,18 +29,18 @@ fn main() {
         Domain::new(grid_bound, &mut grid_output);
 
     // Fill in with IC values (use normal dist for spike in the middle)
-    let width_f = grid_bound.bounds[(0, 1)] as f32 + 1.0;
-    let height_f = grid_bound.bounds[(1, 1)] as f32 + 1.0;
-    let sigma_sq: f32 = (width_f / 25.0) * (width_f / 25.0);
+    let width_f = grid_bound.bounds[(0, 1)] as f64 + 1.0;
+    let height_f = grid_bound.bounds[(1, 1)] as f64 + 1.0;
+    let sigma_sq: f64 = (width_f / 25.0) * (width_f / 25.0);
     input_domain.par_modify_access(100).for_each(
         |mut d: DomainChunk<'_, GRID_DIMENSION>| {
             d.coord_iter_mut().for_each(
                 |(world_coord, value_mut): (
                     Coord<GRID_DIMENSION>,
-                    &mut f32,
+                    &mut f64,
                 )| {
-                    let x = (world_coord[0] as f32) - (width_f / 2.0);
-                    let y = (world_coord[1] as f32) - (height_f / 2.0);
+                    let x = (world_coord[0] as f64) - (width_f / 2.0);
+                    let y = (world_coord[1] as f64) - (height_f / 2.0);
                     let r = (x * x + y * y).sqrt();
                     let exp = -r * r / (2.0 * sigma_sq);
                     *value_mut = exp.exp()
